Implement CSV freshness backend (on-visit decision tree)
Adds the read-side foundation for CSV_IMPORT_POLICY.md §7:

- delivery/parser/csv_parser.py: v2 combined-CSV parser. Validates the
  preamble (line-1 date, line-4 DELIVERY_WORKSHEET_V<n> marker), parses
  the 18 data columns, groups items by supplier alphabetically, and
  produces the same {header, supplier_blocks} shape as PDFWorksheetParser
  so delivery_service consumes either format unchanged. Exports a fast
  read_preamble() helper used by the picker without a full parse.
- delivery/services/csv_picker.py: lists incoming-v2/, partial-reads the
  first ~1KB of each blob, picks the winner per §4 (latest line-1 date,
  tiebreak by line-2 generation timestamp, then Cloud Storage updated).
  Extension-agnostic — IT sometimes uploads without .csv.
- delivery/services/csv_snapshot_service.py: writes the parsed CSV as an
  immutable delivery_csv_snapshots/{YYYY-MM-DD} doc keyed by item_id,
  and provides is_in_process() — the §5 "any deviation from original"
  diff against the current delivery.
- delivery_service: CSV parser wired in, LineItem.id now uses CSV item_id
  when present (per format doc), new consume_csv() orchestrator that
  parses + persists + snapshots in one step.
- delivery/api/routes/csv_freshness.py: GET /api/v1/csv-freshness returns
  the §7 decision tree result (action + csv + loaded). POST /csv-consume
  downloads + loads a CSV at a given source_path.
- delivery/models.py: quantity_expected, expected_cases, and
  total_cases_expected promoted from int to float per the format doc's
  "accept floats; display as-is" decision.

+++ delivery/models.py
@@ -54,7 +54,7 @@ class SupplierStatus(str, Enum):
 
 class LineItemBase(BaseModel):
     """Core line item fields from parsing."""
-    quantity_expected: int = Field(ge=0)
+    quantity_expected: float = Field(ge=0)
     category: str
     raw_description: str
     product_type: str = ""
@@ -101,7 +101,7 @@ class SetPullRequest(BaseModel):
 class SupplierEntryBase(BaseModel):
     """Core supplier entry fields."""
     supplier_name: str
-    expected_cases: int
+    expected_cases: float
     block_sequence: int
 
 
@@ -121,7 +121,7 @@ class DeliveryHeaderModel(BaseModel):
     delivery_date: Optional[date] = None
     day_number: Optional[int] = None
     week_number: Optional[int] = None
-    total_cases_expected: int = 0
+    total_cases_expected: float = 0
 
 
 class DeliveryBase(BaseModel):
@@ -130,7 +130,7 @@ class DeliveryBase(BaseModel):
     delivery_date: Optional[date] = None
     day_number: Optional[int] = None
     week_number: Optional[int] = None
-    total_cases_expected: int = 0
+    total_cases_expected: float = 0
     source_filename: str = ""
     firebase_path: Optional[str] = None
 
@@ -152,7 +152,7 @@ class DeliverySummary(BaseModel):
     id: str
     day_of_week: str
     delivery_date: Optional[date]
-    total_cases_expected: int
+    total_cases_expected: float
     source_filename: str
     status: DeliveryStatus
     parsed_at: Optional[datetime]
